+++ src/main/java/bowling/helper/ValidInputHelper.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.helper;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
