+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
